Adding Multithreading
The python GIL got in the way of the multithreading attempt in the original source.  Using contexts from the multiprocessing library to run in parallel across available CPU cores. Doing this required the objects to be picklable, so moved the locally defined f() function definitions to their own class, and fixed the recursive calls in the SDF getattribute checks.